refactor(runtime): extract computer-use wire codec and contract types (#1405)
